Delete Model and create separate files for each object
