+++ VineyardInternal/DatabaseServiceProtocol.swift
@@ -1,0 +1,8 @@
+//
+//  DatabaseServiceProtocol.swift
+//  VineyardInternal
+//
+//  Created by Vishnesh Jayanthi Ramanathan on 15/09/24.
+//
+
+import Foundation
