+++ VineyardInternal/FirebaseDataManager.swift
@@ -1,0 +1,8 @@
+//
+//  FirebaseDataManager.swift
+//  VineyardInternal
+//
+//  Created by Vishnesh Jayanthi Ramanathan on 16/09/24.
+//
+
+import Foundation
