+++ VineyardInternal/Resolution.swift
@@ -1,0 +1,38 @@
+//
+//  Resolution.swift
+//  VineyardInternal
+//
+//  Created by Vishnesh Jayanthi Ramanathan on 16/09/24.
+//
+
+import Foundation
+
+enum TimeBound: String, CaseIterable {
+    case day
+    case week
+    case month
+}
+
+enum SuccessCheckoff: String{
+    case incomplete
+    case complete
+}
+
+class Resolution: Identifiable {
+    var id = UUID().uuidString
+    var timeBound: TimeBound
+    var name: String
+    var successCheckoff: SuccessCheckoff
+    var progress: Float
+    var goal: Float
+    var freq: Int // Discuss the ways to make this work
+
+    init(timeBound: TimeBound, name: String, successCheckoff: SuccessCheckoff, progress: Float, goal: Float, freq: Int) {
+        self.timeBound = timeBound
+        self.name = name
+        self.successCheckoff = successCheckoff
+        self.progress = progress
+        self.goal = goal
+        self.freq = freq
+    }
+}
